Rename buffer.h and add const
